+++ src/imports/import-grouping/index.ts
@@ -1,8 +1,8 @@
-export * from './ImportGroup';
-export * from './ImportGroupKeyword';
-export * from './ImportGroupOrder';
-export * from './KeywordImportGroup';
-export * from './RegexImportGroup';
-export * from './RemainImportGroup';
-export * from './ImportGroupSettingParser';
-export * from './ImportGroupIdentifierInvalidError';
+export * from './import-group';
+export * from './import-group-identifier-invalid-error';
+export * from './import-group-keyword';
+export * from './import-group-order';
+export * from './import-group-setting-parser';
+export * from './keyword-import-group';
+export * from './regex-import-group';
+export * from './remain-import-group';
