+++ src/imports/import-grouping/remain-import-group.ts
@@ -1,8 +1,8 @@
 import { Import, StringImport } from 'typescript-parser';
 
 import { importSort } from '../../utilities/utility-functions';
-import { ImportGroup } from './ImportGroup';
-import { ImportGroupOrder } from './ImportGroupOrder';
+import { ImportGroup } from './import-group';
+import { ImportGroupOrder } from './import-group-order';
 
 /**
  * Importgroup that processes all imports. Should be used if other groups don't process the import.
