+++ test/tests/imports/import-grouping/RemainImportGroup.test.ts
@@ -12,62 +12,62 @@ chai.should();
 
 describe('RemainImportGroup', () => {
 
-    const rootPath = workspace.workspaceFolders![0].uri.fsPath;
-    let file: File;
-    let importGroup: RemainImportGroup;
-    let generator: TypescriptCodeGenerator;
+  const rootPath = workspace.workspaceFolders![0].uri.fsPath;
+  let file: File;
+  let importGroup: RemainImportGroup;
+  let generator: TypescriptCodeGenerator;
 
-    before(async () => {
-        const parser = Container.get<TypescriptParser>(iocSymbols.typescriptParser);
-        generator = Container.get<TypescriptCodeGeneratorFactory>(iocSymbols.generatorFactory)();
-        file = await parser.parseFile(
-            join(
-                rootPath,
-                'extension/import-grouping/imports.ts',
-            ),
-            rootPath,
-        );
-    });
+  before(async () => {
+    const parser = Container.get<TypescriptParser>(iocSymbols.typescriptParser);
+    generator = Container.get<TypescriptCodeGeneratorFactory>(iocSymbols.generatorFactory)();
+    file = await parser.parseFile(
+      join(
+        rootPath,
+        'extension/import-grouping/imports.ts',
+      ),
+      rootPath,
+    );
+  });
 
-    beforeEach(() => {
-        importGroup = new RemainImportGroup();
-    });
+  beforeEach(() => {
+    importGroup = new RemainImportGroup();
+  });
 
-    it('should process all imports', () => {
-        file.imports.map(i => importGroup.processImport(i)).should.deep.equal([true, true, true, true, true, true]);
-    });
+  it('should process all imports', () => {
+    file.imports.map(i => importGroup.processImport(i)).should.deep.equal([true, true, true, true, true, true]);
+  });
 
-    it('should generate the correct typescript (asc)', () => {
-        for (const imp of file.imports) {
-            if (importGroup.processImport(imp)) {
-                continue;
-            }
-        }
-        generator.generate(importGroup as any).should.equal(
-            `import './workspaceSideEffectLib';\n` +
-            `import 'sideEffectLib';\n` +
-            `import { AnotherFoobar } from './anotherFile';\n` +
-            `import { Foobar } from './myFile';\n` +
-            `import { AnotherModuleFoo } from 'anotherLib';\n` +
-            `import { ModuleFoobar } from 'myLib';\n`,
-        );
-    });
+  it('should generate the correct typescript (asc)', () => {
+    for (const imp of file.imports) {
+      if (importGroup.processImport(imp)) {
+        continue;
+      }
+    }
+    generator.generate(importGroup as any).should.equal(
+      `import './workspaceSideEffectLib';\n` +
+      `import 'sideEffectLib';\n` +
+      `import { AnotherFoobar } from './anotherFile';\n` +
+      `import { Foobar } from './myFile';\n` +
+      `import { AnotherModuleFoo } from 'anotherLib';\n` +
+      `import { ModuleFoobar } from 'myLib';\n`,
+    );
+  });
 
-    it('should generate the correct typescript (desc)', () => {
-        (importGroup as any).order = 'desc';
-        for (const imp of file.imports) {
-            if (importGroup.processImport(imp)) {
-                continue;
-            }
-        }
-        generator.generate(importGroup as any).should.equal(
-            `import 'sideEffectLib';\n` +
-            `import './workspaceSideEffectLib';\n` +
-            `import { ModuleFoobar } from 'myLib';\n` +
-            `import { AnotherModuleFoo } from 'anotherLib';\n` +
-            `import { Foobar } from './myFile';\n` +
-            `import { AnotherFoobar } from './anotherFile';\n`,
-        );
-    });
+  it('should generate the correct typescript (desc)', () => {
+    (importGroup as any).order = 'desc';
+    for (const imp of file.imports) {
+      if (importGroup.processImport(imp)) {
+        continue;
+      }
+    }
+    generator.generate(importGroup as any).should.equal(
+      `import 'sideEffectLib';\n` +
+      `import './workspaceSideEffectLib';\n` +
+      `import { ModuleFoobar } from 'myLib';\n` +
+      `import { AnotherModuleFoo } from 'anotherLib';\n` +
+      `import { Foobar } from './myFile';\n` +
+      `import { AnotherFoobar } from './anotherFile';\n`,
+    );
+  });
 
 });
